Move all response helper function to own class
